refactor: delete dead PromptAffinityStore interface, move types to sql store
The old JSON-based PromptAffinityStore class was already deleted during
the SQL migration. This removes the now-unused interface file and moves
the IPromptAffinityStore and PromptAffinityRecord types directly into
the sql-prompt-affinity-store module to eliminate the unnecessary
indirection.
