+++ src/lib/db/sql-prompt-affinity-store.ts
@@ -1,11 +1,30 @@
 import type { Sql } from "./index.js";
-import type { PromptAffinityRecord } from "../prompt-affinity-store.js";
 import {
   SELECT_ALL_PROMPT_AFFINITY,
   UPSERT_PROMPT_AFFINITY,
   DELETE_PROMPT_AFFINITY,
   UPSERT_PROMPT_AFFINITY_PROMOTE_PROVISIONAL,
 } from "./schema.js";
+
+export interface IPromptAffinityStore {
+  init?(): Promise<void>;
+  warmup?(): Promise<void>;
+  close(): Promise<void>;
+  get(promptCacheKey: string): Promise<PromptAffinityRecord | undefined>;
+  upsert(promptCacheKey: string, providerId: string, accountId: string): Promise<void>;
+  noteSuccess(promptCacheKey: string, providerId: string, accountId: string): Promise<void>;
+  delete(promptCacheKey: string): Promise<void>;
+}
+
+export interface PromptAffinityRecord {
+  readonly promptCacheKey: string;
+  readonly providerId: string;
+  readonly accountId: string;
+  readonly provisionalProviderId?: string;
+  readonly provisionalAccountId?: string;
+  readonly provisionalSuccessCount?: number;
+  readonly updatedAt: number;
+}
 
 const PROVISIONAL_PROMOTION_SUCCESS_COUNT = 2;
