+++ src/lib/provider-strategy/routing/types.ts
@@ -4,7 +4,7 @@ import type { AccountHealthStore } from "../../db/account-health-store.js";
 import type { EventStore } from "../../db/event-store.js";
 import type { ProviderCredential } from "../../key-pool.js";
 import type { PolicyEngine } from "../../policy/index.js";
-import type { IPromptAffinityStore } from "../../prompt-affinity-store.js";
+import type { IPromptAffinityStore } from "../../db/sql-prompt-affinity-store.js";
 import type { ProviderRoutePheromoneStore } from "../../provider-route-pheromone-store.js";
 import type { RequestLogStore } from "../../request-log-store.js";
 import type { QuotaMonitor } from "../../quota-monitor.js";
